requirements.txt
fastapi
uvicorn
openai
PyPDF2
python-multipart
pydantic
aiofiles
sentence-transformers
python-dotenv
PyMuPDF
langchain
pytesseract
numpy
faiss-cpu
pillow
requests